==> Lessons/06_Lab_GLMs.Rmd <==
---
title: "6: Part 1 - Generalized Linear Models"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Spring 2022"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: inline
---

## Objectives
1. Answer questions on M5/A5
2. Answer questions on M6 - GLMs
3. Additional comments on videos - t-test
4. Practice more application GLM to real datasets


## Set up
```{r, message = FALSE}
library(tidyverse)
library(agricolae)   

PeterPaul.chem.nutrients <- read.csv("./Data/Processed/NTL-LTER_Lake_Chemistry_Nutrients_PeterPaul_Processed.csv", stringsAsFactors = TRUE)
# Set date to date format
PeterPaul.chem.nutrients$sampledate <- as.Date(PeterPaul.chem.nutrients$sampledate, format = "%Y-%m-%d")

EPAair <- read.csv("./Data/Processed/EPAair_O3_PM25_NC1819_Processed.csv", stringsAsFactors = TRUE)
# Set date to date format
EPAair$Date <- as.Date(EPAair$Date, format = "%Y-%m-%d")


Litter <- read.csv("./Data/Processed/NEON_NIWO_Litter_mass_trap_Processed.csv", stringsAsFactors = TRUE)
# Set date to date format
Litter$collectDate <- as.Date(Litter$collectDate , format = "%Y-%m-%d")

# Set theme
mytheme <- theme_classic(base_size = 14) +
  theme(axis.text = element_text(color = "black"), 
        legend.position = "top")
theme_set(mytheme)
```

## T-Test

Continuous response, one categorical explanatory variable with two categories (or comparison to a single value if a one-sample test).

### Formulating Hypothesis for µ

Two hypotheses are formed – the null hypothesis and the alternative hypothesis.
The null hypothesis and the alternative hypothesis combine to cover all possible values for the population mean.
The null hypothesis must have the equality.
The null and alternative hypotheses are always stated in terms of the population mean (mu).

### One-sample t-test
The object of a one sample test is to test the null hypothesis that the mean of the group is equal to a specific value. For example, we might ask ourselves (from the EPA air quality processed dataset): 

Function t.test()
**x**	a (non-empty) numeric vector of data values.
**alternative** a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.
**mu** a number indicating the true value of the mean (or difference in means if you are performing a two sample test).
**formula** a formula of the form lhs ~ rhs where lhs is a numeric variable giving the data values and rhs either 1 for a one-sample or paired test or a factor with two levels giving the corresponding groups. If lhs is of class "Pair" and rhs is 1, a paired test is done.


Are Ozone levels below the threshold for "good" AQI index (0-50)?


> Exercise 1: State the hypotheses for testing mean of AQI index.

> Answer:  H0:mu >= 50
          Ha: mu>50


```{r}


summary(EPAair$Ozone)
EPAair.subsample <- sample_n(EPAair, 5000)
summary(EPAair.subsample)

# Evaluate assumption of normal distribution
shapiro.test((EPAair.subsample$Ozone))
ggplot(EPAair, aes(x = Ozone)) +
  geom_histogram() 
qqnorm(EPAair$Ozone); qqline(EPAair$Ozone)

O3.onesample <- t.test(EPAair$Ozone, mu = 50, alternative = "less")
O3.onesample
# p value less than .05, so we reject the null hypothesis and accept the alternative hypothesis: true mean is less than 50
Ozone.plot <- ggplot(EPAair, aes(x = Ozone)) +
  #geom_density(stat = "count", fill = "gray") +
  geom_density(fill = "gray") +
  geom_vline(xintercept = 50, color = "#238b45", lty = 2, size = 0.9) +
  scale_x_continuous(expand = c(0, 0)) + scale_y_continuous(expand = c(0, 0))
print(Ozone.plot)
```

Write a sentence or two about the results of this test. Include both the results of the test and an interpretation that puts the findings in context of the research question.

>  p value less than .05, so we reject the null hypothesis and accept the alternative hypothesis: true mean is less than 50 

### Two-sample t-test

The two-sample *t* test is used to test the hypothesis that the mean of two samples is equivalent. Unlike the one-sample tests, a two-sample test requires a second assumption that the variance of the two groups is equivalent. Are Ozone levels different between 2018 and 2019?

```{r}
shapiro.test(EPAair$Ozone[EPAair$Year == 2018])
shapiro.test(EPAair$Ozone[EPAair$Year == 2019])
#two sample t test- can only have two levels on categorical variable 

#p-value less than 0.05 then reject null for 2018 and 2019 i.e. data do not follow normal distribution

#Compare variance using F-test (only)
var.test(EPAair$Ozone ~ EPAair$Year)

#p-value less than 0.05 then reject null for 2018 and 2019 i.e. true ratio not equal to one

ggplot(EPAair, aes(x = Ozone, color = as.factor(Year))) +
  geom_freqpoly()

# Format as a t-test
O3.twosample <- t.test(EPAair$Ozone ~ EPAair$Year, var.equal = TRUE) #forcing the variance to make same as an anova
O3.twosample
O3.twosample$p.value #just provides the p value

#reject the null: The true difference of means are not the same


# Format as a GLM
O3.twosample2 <- lm(EPAair$Ozone ~ EPAair$Year)
summary(O3.twosample2)

plot(O3.twosample2)


```
when we do anova, we assume variance is the same for all the levels. This assumption is not always the default for t tests, so we test it with a var.test.

## Statistical Test: Cheat sheet

**F-test:** Compare the variances of two groups. The data must be normally distributed.

**Bartlett’s test:** Compare the variances of two or more groups. The data must be normally distributed.

**Shapiro.test:** check for normality

**One-sample t-test:** check if mean is equal/less/greater to specific value, single variable

**Two-sample t-test:**  check if mean of two samples is equivalent

### Visualization and interpretation challenge

Create three plots, each with appropriately formatted axes and legends. Choose a non-default color palette.

1. geom_density of ozone divided by year (distinguish between years by adding transparency to the geom_density layer).
2. geom_boxplot of ozone divided by year . Add letters representing a significant difference between 2018 and 2019 (hint: stat_summary). 
3. geom_violin of ozone divided by year, with the 0.5 quantile marked as a horizontal line. Add letters representing a significant difference between 2018 and 2019. 

```{r}
#Exercise 2:
#number1 geom density
ggplot(EPAair, aes(x = Ozone, fill= as.factor(Year))) + #fill used to fill entire curve, must use as factor function to separate years 2018 and 2019
  geom_density(alpha =.5)+ #alpha =.5 for transparency
  scale_fill_manual(values = c("darkgreen", "darkblue"))+
  labs(x="Ozone AQI value", y="Density", fill="")

#number 2 box plot
ggplot(EPAair, aes(x=as.factor(Year), y=Ozone))+
  stat_summary(geom="text", fun.y = max, vjust=-1, size =4, label = c("b", "a"))+ #size =4 creates size of the text, creating labels for years
  labs(x="", y = "Ozone AQI value")+
  ylim(0,150) +#extending limits so we can see a at the top of the 2019 plot
  geom_boxplot()

ggplot(EPAair, aes(x=as.factor(Year), y=Ozone))+
   stat_summary(geom="text", fun.y = max, vjust=-1, size =4, label = c("b", "a"))+ #size =4 creates size of the text, creating labels for years
  labs(x="", y = "Ozone AQI value")+
  ylim(0,150) +
  geom_violin(draw_quantiles = .5) #marking .5 quantile as horizontal line


  
```

## Linear Regression

Important components of the linear regression are the correlation and the R-squared value. The **correlation** is a number between -1 and 1, describing the relationship between the variables. Correlations close to -1 represent strong negative correlations, correlations close to zero represent weak correlations, and correlations close to 1 represent strong positive correlations. The **R-squared value** is the correlation squared, becoming a number between 0 and 1. The R-squared value describes the percent of variance accounted for by the explanatory variables. 

For the NTL-LTER dataset, can we predict PM2.5 from Ozone?

```{r}

#Exercise 3: Run a linear regression PM2.5 by Ozone. Find the p-value and R-squared value. 
epa.lm<- lm(data = EPAair, PM2.5 ~ Ozone)
summary(epa.lm)

#Exercise 4: Build a scatterplot. Add a line and standard error for the linear regression. 

epaplot.4<-
  ggplot(EPAair, aes(x = Ozone, y = PM2.5)) + #plotting PM2.5 by Ozone
    geom_point()+ #making graph a scatterplot
   geom_smooth(method=lm, color = "blue", se =T) #adding blue line

print(epaplot.4)

```
conclude that we can't rely on ozone to predict PM2.5. there is not a strong enough correlation - we need more variables. 

## AIC to select variables

What other variables can we add to improve model?

```{r}

#Exercise 5: Build correlation plots and identify more possible explanatory variables to add to the regression.

library(corrplot)
EPAair.numeric <- 
 EPAair %>%
  select(meanLat:Ozone) %>%
  na.omit()

EPAcorr <- cor(EPAair.numeric)
corrplot(EPAcorr, method = "ellipse")
corrplot.mixed(EPAcorr, upper = "ellipse")


#appears that lat, long, month and ozone have correlation with PM2.5


#Exercise 6: Choose a model by AIC in a Stepwise Algorithm. Do the results from AIC match the variables you selected on Exercise 5?


EPAAIC <- lm(data = EPAair.numeric, PM2.5 ~ meanLat + meanLng +
              Month + Ozone)
#making step function
step(EPAAIC)

#AIC matches the variables i chose, none should be removed or else the AIC would increase.

#Exercise 7: Run another regression using the variables selected on Exercise 6. Compare r-squared value with the one from Exercise 5. 

epa.multiplelm<- lm(data = EPAair, PM2.5 ~ Ozone + meanLat + meanLng + Month)
summary(epa.multiplelm)

#the multiple r squared is much higher in the new model, meaning we can explain a higher percentage of variance in the PM2.5

```
